CyberOS_rev11/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
whitenoise==6.4.0
Pillow==10.0.0
django-crispy-forms==1.14.0
django-filter==23.2.0